+++ instrumentation/opentelemetry-instrumentation-pyramid/src/opentelemetry/instrumentation/pyramid/__init__.py
@@ -144,7 +144,7 @@ Note:
 API
 ---
 """
-
+import platform
 from typing import Collection
 
 from pyramid.config import Configurator
@@ -166,6 +166,11 @@ from opentelemetry.instrumentation.utils import unwrap
 # from importing an unused symbol.
 trace_tween_factory  # pylint: disable=pointless-statement
 
+if platform.python_implementation() == "PyPy":
+    CALLER_LEVELS = 3
+else:
+    CALLER_LEVELS = 2
+
 
 def _traced_init(wrapped, instance, args, kwargs):
     settings = kwargs.get("settings", {})
@@ -185,10 +190,12 @@ def _traced_init(wrapped, instance, args, kwargs):
     # to find the calling package. So if we let the original `__init__`
     # function call it, our wrapper will mess things up.
     if not kwargs.get("package", None):
-        # Get the package for the third frame up from this one.
-        # Default is `level=2` which will give us the package from `wrapt`
-        # instead of the desired package (the caller)
-        kwargs["package"] = caller_package(level=3)
+        # Get the package for the 2nd frame up from this one.
+        # Default is `level=2` one level down (in Configurator.__init__).
+        # We want the 3rd level from _there_. Since we are already 1 level above,
+        # we need the 2nd level up from here, which will give us the package from
+        # `wrapt` instead of the desired package (the caller)
+        kwargs["package"] = caller_package(level=CALLER_LEVELS)
 
     wrapped(*args, **kwargs)
     instance.include("opentelemetry.instrumentation.pyramid.callbacks")